# comments/pkg/api/v1/comments.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "comments.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Comments"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/comment/add": {
      "post": {
        "operationId": "Comments_CommentAdd",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CommentAddResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentAddRequest"
            }
          }
        ],
        "tags": [
          "Comments"
        ]
      }
    },
    "/comment/edit": {
      "post": {
        "operationId": "Comments_CommentEdit",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CommentEditResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentEditRequest"
            }
          }
        ],
        "tags": [
          "Comments"
        ]
      }
    },
    "/comment/get-by-id": {
      "post": {
        "operationId": "Comments_CommentGetByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CommentGetByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentGetByIDRequest"
            }
          }
        ],
        "tags": [
          "Comments"
        ]
      }
    },
    "/comment/list-by-sku": {
      "post": {
        "operationId": "Comments_CommentListBySku",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CommentListBySKUResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentListBySKURequest"
            }
          }
        ],
        "tags": [
          "Comments"
        ]
      }
    },
    "/comment/list-by-user": {
      "post": {
        "operationId": "Comments_CommentListByUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CommentListByUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentListByUserRequest"
            }
          }
        ],
        "tags": [
          "Comments"
        ]
      }
    }
  },
  "definitions": {
    "CommentAddRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "format": "int64"
        },
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "comment": {
          "type": "string"
        }
      }
    },
    "CommentAddResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "CommentBySku": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "comment": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CommentByUserID": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "comment": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CommentEditRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "format": "int64"
        },
        "comment_id": {
          "type": "string",
          "format": "int64"
        },
        "new_comment": {
          "type": "string"
        }
      }
    },
    "CommentEditResponse": {
      "type": "object"
    },
    "CommentGetByIDRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "CommentGetByIDResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "comment": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CommentListBySKURequest": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "cursor": {
          "$ref": "#/definitions/Cursor"
        }
      }
    },
    "CommentListBySKUResponse": {
      "type": "object",
      "properties": {
        "comments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentBySku"
          }
        },
        "cursor": {
          "$ref": "#/definitions/Cursor"
        }
      }
    },
    "CommentListByUserRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "CommentListByUserResponse": {
      "type": "object",
      "properties": {
        "comments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentByUserID"
          }
        }
      }
    },
    "Cursor": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}